+++ src/index.ts
@@ -1,16 +1,17 @@
-export { Box } from '~/components/Box';
-export { Text } from '~/components/Text';
-export { Button } from '~/components/Button';
-export { IconButton } from '~/components/IconButton';
-export { Icon, IconNames } from '~/components/Icon';
-export { Pre } from '~/components/Code';
+export { Box } from '@components/Box';
+export { Text } from '@components/Text';
+export { Button } from '@components/Button';
+export { IconButton } from '@components/IconButton';
+export { Icon, IconNames } from '@components/Icon';
+export { Pre } from '@components/Code';
 export { ThemeProvider } from '~/contexts/ThemeContext';
-export { ThemeSwitcher } from '~/components/ThemeSwitcher';
-export { Heading } from '~/components/Heading';
-export { Link } from '~/components/Link';
-export { Spinner } from '~/components/Spinner';
-export { Divider } from '~/components/Divider';
-export { CheckBox } from '~/components/CheckBox';
-export { Radio } from '~/components/Radio';
-export { TextInput } from '~/components/TextInput';
-export { Textarea } from '~/components/Textarea';
+export { ThemeSwitcher } from '@components/ThemeSwitcher';
+export { Heading } from '@components/Heading';
+export { Link } from '@components/Link';
+export { Spinner } from '@components/Spinner';
+export { Divider } from '@components/Divider';
+export { CheckBox } from '@components/CheckBox';
+export { Radio } from '@components/Radio';
+export { TextInput } from '@components/TextInput';
+export { Textarea } from '@components/Textarea';
+export * as CetecPreset from '../panda-preset';
